Resolve defect in PATENTS.TXT
Rationale: The "as part of" conditions don't make sense, rely on an impossible-to-evaluate ".NET Runtime" definition, and - I'm told by people who should know - were not intended to limit the scope of the license. 

Based on that information, I suggest the removal of any mention of the .NET Runtime, and the replacement of these unintentional conditions with an an unambiguous "Covered Code, a portion thereof, or a derivative work thereof.", which should cover all use cases that the original version intended (but failed) to include with the "as  part of" phrasing. 

Fixes #375